Cleanup: Split types/index.ts by domain (config, IR, expressions) [S] (#268)
